merge upstream

+++ kiwi/utils/rpm_database.py
@@ -24,7 +24,7 @@ from kiwi.path import Path
 from kiwi.utils.rpm import Rpm
 
 
-class RpmDataBase(object):
+class RpmDataBase:
     """
     **Setup RPM database configuration**
     """
@@ -35,13 +35,11 @@ class RpmDataBase(object):
 
     def has_rpm(self):
         """
-        Check if rpm binary was found in root_dir
+        Check if rpmdb binary was found in root_dir to indicate
+        that the rpm system is present.
         """
-        rpm_search_env = {
-            'PATH': os.sep.join([self.root_dir, 'usr', 'bin'])
-        }
         rpm_bin = Path.which(
-            'rpm', custom_env=rpm_search_env, access_mode=os.X_OK
+            'rpmdb', root_dir=self.root_dir, access_mode=os.X_OK
         )
         if not rpm_bin:
             return False
@@ -107,6 +105,33 @@ class RpmDataBase(object):
         )
         self.rpmdb_image.write_config()
 
+    def link_database_to_host_path(self):
+        """
+        Create a link from host database location to image database
+        location. The link is only created if both locations differ
+        and if the host database location does not exist.
+        """
+        rpm_host_dbpath = self.rpmdb_host.expand_query('%_dbpath')
+        rpm_image_dbpath = self.rpmdb_image.expand_query('%_dbpath')
+        if rpm_host_dbpath != rpm_image_dbpath:
+            root_rpm_host_dbpath = os.path.normpath(
+                os.sep.join([self.root_dir, rpm_host_dbpath])
+            )
+            if not os.path.exists(root_rpm_host_dbpath):
+                Path.create(os.path.dirname(root_rpm_host_dbpath))
+                host_to_root = ''.join(
+                    '../' for i in os.path.dirname(
+                        rpm_host_dbpath
+                    ).lstrip(os.sep).split(os.sep)
+                )
+                Command.run(
+                    [
+                        'ln', '-s', os.path.normpath(
+                            os.sep.join([host_to_root, rpm_image_dbpath])
+                        ), root_rpm_host_dbpath
+                    ]
+                )
+
     def set_database_to_image_path(self):
         """
         Setup dbpath to point to the image rpm dbpath configuration
@@ -117,26 +142,33 @@ class RpmDataBase(object):
         rpm_image_dbpath = self.rpmdb_image.expand_query('%_dbpath')
         rpm_host_dbpath = self.rpmdb_host.expand_query('%_dbpath')
         if rpm_image_dbpath != rpm_host_dbpath:
+            root_rpm_image_dbpath = os.path.normpath(
+                os.sep.join([self.root_dir, rpm_image_dbpath])
+            )
+            root_rpm_host_dbpath = os.path.normpath(
+                os.sep.join([self.root_dir, rpm_host_dbpath])
+            )
+
+            if os.path.islink(root_rpm_host_dbpath):
+                self.rebuild_database()
+                return
+
+            if os.path.islink(root_rpm_image_dbpath):
+                os.unlink(root_rpm_image_dbpath)
+            else:
+                Path.wipe(root_rpm_image_dbpath)
+
             self.rpmdb_image.set_config_value(
                 '_dbpath', rpm_host_dbpath
             )
             self.rpmdb_image.set_config_value(
                 '_dbpath_rebuild', self.rpmdb_image.get_query('_dbpath')
             )
-            root_rpm_image_dbpath = os.path.normpath(
-                os.sep.join([self.root_dir, rpm_image_dbpath])
-            )
-            if os.path.islink(root_rpm_image_dbpath):
-                os.unlink(root_rpm_image_dbpath)
-            else:
-                Path.wipe(root_rpm_image_dbpath)
+
             self.rpmdb_image.write_config()
             self.rebuild_database()
             self.rpmdb_image.wipe_config()
 
-            root_rpm_host_dbpath = os.path.normpath(
-                os.sep.join([self.root_dir, rpm_host_dbpath])
-            )
             root_rpm_alternatives = os.sep.join(
                 [root_rpm_host_dbpath, 'alternatives']
             )
